adds adjust board feature
Now, evaluate() will take a board (after the move has been made), remove each match that the move made, slide down the orbs on top of the match, and generate random orbs to fill in the rest of the board.

NOTE: If new matches occur after the board is adjusted, the board needs to be avaluated again.

NOTE: I did not test what happens if two matches appear on top of each other with the same move, but I believe the logic deals with it correctly as is.