Update match nomenclature to make more sense
hasMatch() becomes hasPotentialMatch()
hasMatchEvent() becomes hasMatch()
findMatches() becomes findTriples()
foundMatches becomes triples

I found these names really confusing going back through the code. I believe
that a match is when there are three or more connected orbs on the board. I
also believe a potential match is an instance when a match can be made by making
one swap on the board. This was already reflected in the test descriptions.

Also, findMatches() was not actaully finding matches. It was only finding parts
of matches. I was going to make it something like findConsecutiveThreeInstances(),
but I think triples works nice and neatly.