Add atticOrbs as a default attribute in Board
When a Board is initiated, by default `this.atticOrbs` will
be a new set of orbs the same size and with the same types as the main
set of orbs. The `atticOrbs` will also not have a match already and will
have at least one possible match. This is done by creating a new Board instance
within the constructor, and assigning that instance's `this.orbs` to the
`atticOrbs`. If `atticOrbs` are given as an argument when the board is initiated,
that argument will be assigned directly to `this.atticOrbs`

This addition prepares a new concept for replacing orbs in `evaluate()`.
The new concept will place a separate set of orbs 'above' the main
boards. Once the match has been made and the orbs 'deleted', orbs
from the `atticOrbs` will be dropped down to fill in the empty spaces.