Fix uncaught error in board.swap()
Somehow, the board.swap() function was not changing the orbs in
the demo Board instance. All the tests pass both ways.

I will look more closely at the tests for swap to make sure they
are certainly reliable, but this edit is what should be there, and
it works for the demo.