feat(demo): Improve functionality to handle multiple moves
The demo now works well! But of course there are a couple bugs that
haven't been dealt with yet. Try to figure them out yourself if you
dare! Or just look down a couple lines...

Now that evaluate has been updated on the back end, after the move
in the demo is complete, the board and the attic are repopulated
accurately on in the html.

- animate.createHTMLBoard has been edited to only take (orbs) as a
parameter since it will now only create the main board. The attic
has its own create function which is animate.createHTMLAttic()

- Problems still to be dealt with:
    1) The demo only does one evaluation per swap. That means if
    another match is made from the atticOrbs dropping down, it is
    not evaluated until after the next swap.
    2) The demo also doesn't check if it needs a shuffle after the
    move. An impossible board could happen.
    3) Probably more too. Let me know!