feat(demo): Adjust the animations in evaluate to use transition event listeners
Removing the setTimeouts allows the animations to happen one after
another more naturally, rather than having to know the transtion
times of each animation to call in the setTimeout.

There is a strange issue with the opacity transition in the
eraseMatches function. As a quick fix, I kept one setTimeout.

Also, evaluate has had the swap animation removed, and the scoreboard
is updated directly within evaluate.