feat(attic): Move atticOrbs from the constructor to a getter
This stops the constructor from calling tiself from within itself.
Instead, whenever atticOrbs are referenced, they are created on the
spot. Additionally, there is now an atticTypes parameter in the
constructor that defaults to the previously given types parameter.

This way, at any point, atticTypes can be changed, which will affect
the contents of the attic when they are called for.