Restructure the project to remove babel and incorporate typescript
A number of things are happening to make this all work.
- Removed .babelrc file
- Setup tsconfig.json file
- Replaced all babel compiling scripts with the typescript
transpiling script, 'tsc'.
- Installed 'typings' to manage `.d.ts` files for external libraries
- Upgrade Ava to 0.15.2 to include an `index.d.ts` file
    (some further updating of tests may be in order to match the
    upgrade, but all tests are passing)

The reason that typings was installed and Ava was upgraded have to
do with the module imports in the source code. If a file has a .ts
extension, then when it is required with CommonJS, it will look for
a .ts file as the module to require. In these cases, lodash, and ava
are .js files and therefore could not be imported. The `.d.ts` files
from typings and the newer version of Ava provide typings for these
modules. In theory, the files in node_modules could just be changed
to have the .ts extension, since all Typescript is Javascript, but
that process could be quite tedious. Also, this way, the benefits
of Typescript are extended into the modules themselves. Now, if
an incorrect type of parameter is passed into a lodash function, a
TS error will be thrown.