Update tsc commands to separate test from src
In the typescript compiler, files cannot be excluded if they are
depended upon by files that are included. In this project, that means
compiling the tests will automatically compile the source. That
wouldn't be a problem, except Ava requires typescript to compile to
es2015 in order for the type declaration file to work. However,
we need the source to be compiled to es5.

Now, there are two tsconfig files. `tsconfig.json` is the default,
which will run on the plain `tsc` command. It will only compile the
typescript/src file to es5.

`tsconfig.test.json` can be run by adding the `-p` flag to `tsc` like
this: `tsc -p tsdconfig.test.json`. That will compile the entire
typescript folder into es2015. That command has been put into
`npm run tsc-test`.

Also, `npm test` now runs the tsc-test command, then tsc, then ava.